+++ dsb-integration/src/test/java/com/spectralogic/dsbrowser/integration/CloseConfirmationHandlerTest.java
@@ -50,6 +50,7 @@ import javafx.application.Platform;
 import javafx.collections.ObservableList;
 import javafx.concurrent.Task;
 import javafx.embed.swing.JFXPanel;
+import javafx.util.Pair;
 import org.junit.BeforeClass;
 import org.junit.Test;
 import org.mockito.Mockito;
@@ -78,6 +79,7 @@ public class CloseConfirmationHandlerTest {
     private boolean successFlag = false;
     private final static ResourceBundle resourceBundle = ResourceBundle.getBundle("lang", new Locale(ConfigProperties.getInstance().getLanguage()));
     private static final BuildInfoServiceImpl buildInfoService = new BuildInfoServiceImpl();
+    private static Path path;
 
     @BeforeClass
     public static void setConnection() {
@@ -94,7 +96,7 @@ public class CloseConfirmationHandlerTest {
             session = createConnectionTask.createConnection(SessionModelService.setSessionModel(savedSession, false), resourceBundle, buildInfoService);
             handler = new CloseConfirmationHandler(resourceBundle, jobWorkers, Mockito.mock(ShutdownService.class));
             try {
-                final Path path = ResourceUtils.loadFileResource("files/");
+                path = ResourceUtils.loadFileResource("files/");
                 if (path != null) {
                     file = path.toFile();
                 }
@@ -246,16 +248,16 @@ public class CloseConfirmationHandlerTest {
     public void cancelAllRunningTasks() throws InterruptedException {
         final CountDownLatch latch = new CountDownLatch(1);
         Platform.runLater(() -> {
-            final ImmutableList<File> filesList = ImmutableList.of(file);
+            final ImmutableList<Pair<String,Path>> pair = ImmutableList.of(new Pair<>("files/", path));
             try {
                 final Ds3Client ds3Client = session.getClient();
                 final DeepStorageBrowserPresenter deepStorageBrowserPresenter = Mockito.mock(DeepStorageBrowserPresenter.class);
                 final SettingsStore settingsStore = SettingsStore.loadSettingsStore();
                 final Ds3Common ds3Common = new Ds3Common();
                 ds3Common.setDeepStorageBrowserPresenter(deepStorageBrowserPresenter);
-                final Ds3PutJob ds3PutJob = new Ds3PutJob(ds3Client, filesList, "cancelAllTasksBucket", "",
-                        Priority.URGENT.toString(), 5, JobInterruptionStore.loadJobIds(),
-                        deepStorageBrowserPresenter, session, settingsStore, Mockito.mock(LoggingService.class), resourceBundle);
+                final Ds3PutJob ds3PutJob = new Ds3PutJob(ds3Client, pair, "cancelAllTasksBucket", "",
+                        JobInterruptionStore.loadJobIds(), Priority.URGENT.toString(), 5, resourceBundle,
+                        settingsStore, Mockito.mock(LoggingService.class), deepStorageBrowserPresenter);
                 jobWorkers.execute(ds3PutJob);
                 ds3PutJob.setOnSucceeded(event -> {
                     System.out.println("Put job success");
